3 new Spleef maps (2 normal, 1 run)
Geared-Up Spleef is a map shaped like a gear. I figured it'd be neat as an alternative theme to donut and maybe provide some neat extra bits I didn't think of.
Spleef: To The Moon! is a map shaped like a rocket ship! It has few things blocked off, except a wall to represent the window, and the fire is not enterable. Maybe once more map features come I'll change it to magma blocks and open it up!
Feather-Footed Run is a simple medium-sized TNT Run config shaped vaguely like a feather. Does not loop, so you'll have to be light on your feet as you retrace your steps if you want to stay on top!